refactor: organize index.js functions to different files, revert 6e48ea8edc503b2d04f91947b2861544ec1fefa5
